Enhanced Physics
Added a transfer of momentum from the paddle. The ball can now be accelerated on the y axis depending on the direcction of the bat. If the player is moving up and the ball is moving up when they collide the ball will move up faster. If they are in opposite directions the speed on the Y axis will slow down and can actually reach 0 (The ball will bounce in a straight line on the x-axis until a player adds / subtracts to the y-axis speed).